refactor(configuration): All the configuration directive will be stored in a Directive called MainBlock. The ConfigurationDatabase is now the global Configuration class
