+++ Makefile
@@ -37,11 +37,11 @@ MAIN_SRC:= \
 CONFIGURATION_SRC:= \
 	Configuration.cpp \
 	Configuration/Directive.cpp \
-	Configuration/Database.cpp \
 	Configuration/Cache/LocationQuery.cpp \
 	Configuration/Cache/ServerQuery.cpp \
 	Configuration/Directive/Socket.cpp \
 	Configuration/Directive/Block.cpp \
+	Configuration/Directive/Block/Main.cpp \
 	Configuration/Directive/Block/Http.cpp \
 	Configuration/Directive/Block/Server.cpp \
 	Configuration/Directive/Block/Events.cpp \
